% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadings.R
\name{loadings}
\alias{loadings}
\title{Returns the loadings of the binary linear classifier (e.g. DWD)}
\usage{
loadings(dpp, loadnum = length(dpp$w))
}
\arguments{
\item{dpp}{A DiProPerm list.}

\item{loadnum}{An integer indicating the number of variables to display.  For example, if loadnum=5 then the indexes for the five variables who contributed most toward the separation of the two classes are displayed. The default is to print out all the loadings.}
}
\value{
Returns the indexes and loadings for variables which contributed the most toward the separation of the the binary classifier.
}
\description{
Returns the variable indexes who had the highest loadings in the binary classification.  Higher loading values indicate a variable's contribution toward the separation between the two classes.
}
\examples{
\donttest{data(mushrooms)
X <- Matrix::t(mushrooms$X)
y <- mushrooms$y
dpp <- DiProPerm(X=X,y=y,B=10)
loadings(dpp,loadnum=3)}
\dontshow{data(mushrooms)
X.temp <- SparseM::as.matrix(mushrooms$X)[,1:50]
X <- Matrix::t(X.temp)
y <- mushrooms$y[1:50]
dpp <- DiProPerm(X=X,y=y,B=100,classifier="md")
loadings(dpp,loadnum=3)}
}
\author{
Andrew G. Allmon, J.S. Marron, Michael G. Hudgens
}
